Optimization. Part 3.
- add more noexcept
- add declarations and definitions for default class methods
- add comparison operators for MusicTrack and SoundChunk
- remove unnecessary `static_cast`s in RenderSystem::update
- add get() for SoundChannel
- use const for some constant vars

+++ BloomFramework/include/Systems/RenderSystem.h
@@ -16,12 +16,7 @@ namespace bloom::systems {
 		void update(std::optional<double> deltaTime = std::nullopt) override {
 			m_registry.view<Position, Size, Sprite>().each(
 				[](auto entity, Position & pos, Size & size, Sprite & spr) {
-				SDL_Rect destRect{
-					static_cast<int>(pos.x),
-					static_cast<int>(pos.y),
-					static_cast<int>(size.w),
-					static_cast<int>(size.h)
-				};
+				SDL_Rect destRect{ pos.x, pos.y, size.w, size.h };
 				spr.texture->render(spr.srcRect, destRect);
 			});
 		}
